Make service notification run on server, not locally
This means we no longer require secrets to be stored in development
machines
